GUI: IOtherAppWindow.SendDataMessage(): Use IntPtr for MessageId.
git-svn-id: https://svn.rinrab.com/rinrab/MKW/trunk@5294 0083e6b4-6090-234a-836e-88fcc467669c

+++ src/MKW.GUI/SingleInstance/IOtherAppWindow.cs
@@ -2,6 +2,6 @@
 {
     internal interface IOtherAppWindow
     {
-        bool SendDataMessage(uint messageId, byte[] data);
+        bool SendDataMessage(IntPtr messageId, byte[] data);
     }
 }
